Python/requirements.txt
numpy
Pillow
transformers